管理中心接口统一添加鉴权

+++ src/main/java/com/jfshare/mvp/server/controller/AdminController.java
@@ -32,7 +32,6 @@ import com.jfshare.mvp.server.service.JvjindouRuleService;
 import com.jfshare.mvp.server.service.ProductItemService;
 import com.jfshare.mvp.server.service.PromotionSettingService;
 import com.jfshare.mvp.server.service.SystemInformationService;
-import com.jfshare.mvp.server.utils.JedisClusterUtils;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -71,14 +70,26 @@ public class AdminController {
 	
 	@ApiOperation(value = "IOS上线审核设置", notes = "保存IOS上线审核设置")
 	@PostMapping("/appVerifySetting")
-	public ResultConstant saveAppVerifySetting(@RequestBody TbAppVerifySetting tbAppVerifySetting) {
+	public ResultConstant saveAppVerifySetting(@RequestBody TbAppVerifySetting tbAppVerifySetting,
+			@RequestParam(value = "token", required = true) String token) {
+		//鉴权校验
+		Map<String, Object> mapResult = adminService.checkToken(token);
+		if(!(boolean) mapResult.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,mapResult.get("msg").toString());
+		}
 		return appVerifySettingService.saveAppVerifyProducts(tbAppVerifySetting);
 	}
 	
 	@SuppressWarnings("rawtypes")
 	@ApiOperation(value = "保存并发布推广微页面设置", notes = "保存并发布传入的推广配置和类目商品展示配置")
 	@PostMapping("/promotionSetting")
-	public ResultConstant savePromotionSetting(@RequestBody Map<String, List<Map>> map) {
+	public ResultConstant savePromotionSetting(@RequestBody Map<String, List<Map>> map,
+			@RequestParam(value = "token", required = true) String token	) {
+		//鉴权校验
+		Map<String, Object> mapResult = adminService.checkToken(token);
+		if(!(boolean) mapResult.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,mapResult.get("msg").toString());
+		}
 		List<Map> productPromotions = map.get("productPromotions");
 		List<Map> productItemShows = map.get("productItemShows");
 		boolean result = promotionSettingService.savePromotionSetting(productPromotions, productItemShows);
@@ -134,7 +145,14 @@ public class AdminController {
 
 	@ApiOperation(value = "更新商品类目", notes = "根据传入的商品类目配置，重新配置商品类目")
 	@PostMapping("/productItem")
-	public ResultConstant updateProductItem(@RequestBody TbProductItem tbProductItem) {
+	public ResultConstant updateProductItem(@RequestBody TbProductItem tbProductItem,
+			@RequestParam(value = "token", required = true) String token) {
+		
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
 		boolean result = false;
 		if (StringUtils.isEmpty(tbProductItem.getItemNo())) {
 			result = productItemService.addProductItem(tbProductItem.getItemName(), tbProductItem.getItemDesc(), tbProductItem.getParentItemNo());
@@ -170,7 +188,14 @@ public class AdminController {
 
 	@ApiOperation(value = "删除商品类目", notes = "根据传入的商品类目编号，删除商品类目")
 	@DeleteMapping("/productItem")
-	public ResultConstant deleteProductItem(@RequestBody Map<String, List<String>> itemNos) {
+	public ResultConstant deleteProductItem(@RequestBody Map<String, List<String>> itemNos,
+			@RequestParam(value = "token", required = true) String token) {
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
+		
 		ResultConstant result = productItemService.deleteProductItem(itemNos.get("itemNo"));
 		return result;
 	}
@@ -262,8 +287,13 @@ public class AdminController {
 			@RequestParam(value="givingRule",required=true)String givingRule,
 			@RequestParam(value="randomGivingMin",required=false)Integer randomGivingMin,
 			@RequestParam(value="randomGivingMax",required=false)Integer randomGivingMax,
-			@RequestParam(value="fixedGiving",required=false)Integer fixedGiving
-			) {
+			@RequestParam(value="fixedGiving",required=false)Integer fixedGiving,
+			@RequestParam(value="token",required=true)String token) {
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
 		TbJvjindouRule jvjindouRule =jvjindouRuleService.getJvjindouRule(id);
 		if(givingRule.equals(Constant.FIXED_PATTERN)) {
 			if(!StringUtils.isEmpty(fixedGiving)){
@@ -288,7 +318,14 @@ public class AdminController {
 	}
 	@ApiOperation(value = "聚金豆规则信息添加", notes = "添加聚金豆规则设定")
 	@PostMapping("/addjvjindouRule")
-	public ResultConstant addjvjindouRule(TbJvjindouRule jvjindouRule ) {
+	public ResultConstant addjvjindouRule(TbJvjindouRule jvjindouRule ,
+			@RequestParam(value="token",required=true)String token) {
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
+		
 		if(StringUtils.isEmpty(jvjindouRule)) {
 			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR, "参数有误");
 		}
@@ -322,7 +359,14 @@ public class AdminController {
 	public ResultConstant addInformation(
 			@RequestParam(value="title",required=true)String title,
 			@RequestParam(value="cont",required=true)String cont,
-			@RequestParam(value="user",required=true)String user) {
+			@RequestParam(value="user",required=true)String user,
+			@RequestParam(value="token",required=true)String token) {
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
+		
 		TbSystemInformation systemInformation = new TbSystemInformation();
 		Date date  = new Date();
 		systemInformation.setTitle(title);
@@ -341,7 +385,15 @@ public class AdminController {
 	public ResultConstant addAndReleaseInformation(
 			@RequestParam(value="title",required=true)String title,
 			@RequestParam(value="cont",required=true)String cont,
-			@RequestParam(value="user",required=true)String user) {
+			@RequestParam(value="user",required=true)String user,
+			@RequestParam(value="token",required=true)String token) {
+		
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
+		
 		TbSystemInformation systemInformation = new TbSystemInformation();
 		Date date  = new Date();
 		systemInformation.setTitle(title);
@@ -389,7 +441,13 @@ public class AdminController {
 	public ResultConstant updateInformation(
 			@RequestParam(value="id",required=true)Integer id,
 			@RequestParam(value="title",required=true)String title,
-			@RequestParam(value="cont",required=true)String cont) {
+			@RequestParam(value="cont",required=true)String cont,
+			@RequestParam(value="token",required=true)String token) {
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
 		TbSystemInformation systemInformation = systemInformationService.getInformatinInfo(id);
 		int result=0;
 		if(systemInformation!=null) {
@@ -405,7 +463,13 @@ public class AdminController {
 	
 	@ApiOperation(value="系统消息删除",notes="删除系统消息，id:消息id")
 	@PostMapping("/deleteInformation")
-	public ResultConstant deleteInformation(@RequestParam(value="id",required=true)String id) {
+	public ResultConstant deleteInformation(@RequestParam(value="id",required=true)String id,
+			@RequestParam(value="token",required=true)String token) {
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
 		int result=0;
 		String [] ids = id.split(",");
 		for(int i=0;i<ids.length;i++) {
@@ -432,7 +496,13 @@ public class AdminController {
 	@ApiOperation(value="系统消息发布",notes="系统消息发布，id:消息id")
 	@PostMapping("/releaseformation")
 	public ResultConstant releaseformation(
-			@RequestParam(value="id",required=true)Integer id) {
+			@RequestParam(value="id",required=true)Integer id,
+			@RequestParam(value="token",required=true)String token) {
+		//鉴权校验
+		Map<String, Object> map = adminService.checkToken(token);
+		if(!(boolean) map.get("result")){
+			return ResultConstant.ofFail(ResultConstant.FAIL_CODE_PARAM_ERROR,map.get("msg").toString());
+		}
 		TbSystemInformation systemInformation = systemInformationService.getInformatinInfo(id);
 		int result=0;
 		if(systemInformation!=null && systemInformation.getStatus()==1) {
